lib: split `atomptr_*` out from `atomlist.h`
Just a little reorg to split off unrelated things.
