+++ lib/subdir.am
@@ -202,6 +202,7 @@ nobase_pkginclude_HEADERS += \
 	lib/agg_table.h \
 	lib/asn.h \
 	lib/atomlist.h \
+	lib/atomptr.h \
 	lib/base64.h \
 	lib/bfd.h \
 	lib/bitfield.h \
